requirements.txt
DBUtils==1.2
fastapi
pydantic
pymysql
pyjwt
yt-dlp
requests